+++ pufferlib/config/ocean/drive.ini
@@ -69,7 +69,7 @@ obs_only = True # If True, show exactly what the agent sees in agent observation
 show_grid = False # Show grid lines
 show_lasers = False # Draws lines from ego agent observed ORUs and road elements to show detection range
 show_human_logs = True # Display human xy logs in the background
-render_map = "pufferlib/resources/drive/binaries/map_000.bin" # Options: str to path (e.g., "resources/drive/binaries/map_001.bin"), None
+render_map = none  # Options: str to path (e.g., "resources/drive/binaries/map_001.bin"), None
 
 [sweep.env.reward_vehicle_collision]
 distribution = uniform
